Added new classes and a movement call to the frame.
I added a player and controlls class to the model package. The player
package only has an x and y parameter so far. The rectangle succesfully
moves right. The control class was given a method to listen for keys
pressed and the frame has a controller object to make movement happen.